+++ lib/generators/cms/component/form_builder/templates/example_migration.rb
@@ -1,13 +1,16 @@
 class CreateFormBuilderExample < ::RailsConnector::Migration
-  def path
-    '<%= cms_path %>/form-builder'
-  end
-
   def up
     create_obj(
-      _path: path,
+      _path: '<%= cms_path %>/feedback',
       _obj_class: '<%= class_name %>',
-      title: 'Contact',
+      title: 'Feedback',
+      '<%= title_attribute_name %>' => 'Feedback',
+      '<%= body_attribute_name %>' => 'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed
+        do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
+        nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute
+        irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
+        Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit
+        anim id est laborum.',
       '<%= crm_activity_type_attribute_name %>' => '<%= activity_type %>',
       '<%= show_in_navigation_attribute_name %>' => 'Yes'
     )